[libs-core] rename: BaseBlock -> BaseObject (#3)
