+++ libs/core/src/main/java/kr/joberchip/core/space/block/ImageBlock.java
@@ -4,8 +4,8 @@ import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.OneToOne;
 import javax.persistence.Table;
-import kr.joberchip.core.file.ImageBlockFile;
-import kr.joberchip.core.space.BaseBlock;
+import kr.joberchip.core.storage.ImageBlockFile;
+import kr.joberchip.core.space.BaseObject;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.NoArgsConstructor;
@@ -14,7 +14,7 @@ import lombok.NoArgsConstructor;
 @Table(name = "image_block_tb")
 @NoArgsConstructor
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
-public class ImageBlock extends BaseBlock {
+public class ImageBlock extends BaseObject {
   @Column(name = "title")
   private String title;
