+++ libs/core/src/main/java/kr/joberchip/core/space/block/LinkBlock.java
@@ -3,11 +3,11 @@ package kr.joberchip.core.space.block;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.Table;
-import kr.joberchip.core.space.BaseBlock;
+import kr.joberchip.core.space.BaseObject;
 
 @Entity
 @Table(name = "link_block_tb")
-public class LinkBlock extends BaseBlock {
+public class LinkBlock extends BaseObject {
   @Column(name = "title")
   private String title;
