+++ libs/core/src/main/java/kr/joberchip/core/space/block/TemplateBlock.java
@@ -3,7 +3,7 @@ package kr.joberchip.core.space.block;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.Table;
-import kr.joberchip.core.space.BaseBlock;
+import kr.joberchip.core.space.BaseObject;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.NoArgsConstructor;
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 @Table(name = "template_block_tb")
 @NoArgsConstructor
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
-public class TemplateBlock extends BaseBlock {
+public class TemplateBlock extends BaseObject {
   @Column(name = "title")
   private String title;
