+++ libs/core/src/main/java/kr/joberchip/core/space/block/TextBlock.java
@@ -4,11 +4,11 @@ import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.Lob;
 import javax.persistence.Table;
-import kr.joberchip.core.space.BaseBlock;
+import kr.joberchip.core.space.BaseObject;
 
 @Entity
 @Table(name = "text_block_tb")
-public class TextBlock extends BaseBlock {
+public class TextBlock extends BaseObject {
   @Lob
   @Column(name = "content")
   private String content;
